+++ tests/test-dirs/destruct/exceptions_pre_408.t
@@ -1,6 +1,6 @@
 (enabled_if (< %{ocaml_version} 4.08.0))
 FIXME
-  $ $MERLIN single case-analysis -start 3:4 -end 3:8 -filename complete.ml -log-file /tmp/mlog2 <<EOF \
+  $ $MERLIN single case-analysis -start 3:4 -end 3:8 -filename complete.ml <<EOF \
   > let _ = \
   >   match (None : int option) with \
   >   | exception _ -> () \
@@ -25,12 +25,37 @@ FIXME
     "notifications": []
   }
 
-FIXME
-  $ $MERLIN single case-analysis -start 4:4 -end 4:8 -filename complete.ml -log-file /tmp/mlog2 <<EOF \
+  $ $MERLIN single case-analysis -start 4:4 -end 4:8 -filename complete.ml <<EOF \
   > let _ = \
   >   match (None : int option) with \
   >   | exception _ -> () \
   >   | Some _ -> () \
+  > EOF
+  {
+    "class": "error",
+    "value": "Nothing to do",
+    "notifications": []
+  }
+
+  $ $MERLIN single case-analysis -start 4:5 -end 4:5 -filename no_comp_pat.ml <<EOF \
+  > let _ = \
+  >   match (None : unit option) with \
+  >   | exception _ -> () \
+  >   | None -> () \
+  > EOF
+  {
+    "class": "error",
+    "value": "Nothing to do",
+    "notifications": []
+  }
+
+FIXME: `Some 0` certainly is a missing case but we can do better:
+
+  $ $MERLIN single case-analysis -start 4:4 -end 4:8 -filename complete.ml <<EOF \
+  > let _ = \
+  >   match (None : int option) with \
+  >   | exception _ -> () \
+  >   | Some 3 -> () \
   > EOF
   {
     "class": "error",